refactor(terminal): centralize pane layout definitions
Introduce a dedicated terminal layout registry for the current canonical layouts, move keyboard cycle and layout legality checks onto that shared source, and keep SplitView's layouts module as a compatibility shim while behavior stays unchanged.
