+++ src/features/terminal/layout-registry/index.ts
@@ -1,0 +1,11 @@
+export {
+  LAYOUT_CYCLE,
+  LAYOUTS,
+  VISIBLE_LAYOUTS,
+  autoShrinkLayoutFor,
+  isKnownLayoutId,
+} from './layoutRegistry'
+
+export { LAYOUT_IDS } from './layoutIds'
+
+export { LAYOUT_SPECS, type LayoutShape } from './layoutSpecs'
